can add items to cart for non-signed in user
