+++ client/store/Cart.js
@@ -17,24 +17,17 @@ export const addCart = (id) => async (dispatch) => {
 };
 
 const initialState = [];
-let coffeeIndex;
+let coffeeIndex, tempProduct;
 
 export default function addCartReducer(state = initialState, action) {
   switch (action.type) {
     case ADD_CART:
-      if (state.length) {
-        coffeeIndex = state.findIndex(
-          (coffee) => coffee.id === action.coffee.id
-        );
-        console.log(">>>>>", state);
-        console.log("****", action.coffee.id);
-        if (coffeeIndex >= 0) {
-          state[coffeeIndex].cartQuantity += 1;
-        }
-      } else {
-        console.log(state);
-        return [...state, { ...action, cartQuantity: 1 }];
-      }
+      tempProduct = { ...action, cartQuantity: 1 };
+      coffeeIndex = state.findIndex(
+        (item) => item.coffee.id === action.coffee.id
+      );
+      if (coffeeIndex >= 0) state[coffeeIndex].cartQuantity += 1;
+      else return [...state, tempProduct];
       return state;
     default:
       return state;
